refactor: application 패키지 생성
