+++ src/main/java/org/store/clothstar/member/application/SellerServiceApplication.java
@@ -1,9 +1,10 @@
-package org.store.clothstar.member.service;
+package org.store.clothstar.member.application;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.store.clothstar.member.domain.Seller;
 import org.store.clothstar.member.dto.request.CreateSellerRequest;
+import org.store.clothstar.member.service.SellerBasicService;
 
 @Service
 @RequiredArgsConstructor
