+++ src/test/java/org/store/clothstar/member/service/MemberServiceApplicationMockUnitTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.store.clothstar.member.application.MemberServiceApplication;
 import org.store.clothstar.member.domain.Member;
 import org.store.clothstar.member.dto.response.MemberResponse;
 import org.store.clothstar.member.repository.MemberMybatisRepository;
