+++ src/test/java/org/store/clothstar/member/service/MemberServiceApplicationUnitTest.java
@@ -5,6 +5,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
+import org.store.clothstar.member.application.MemberServiceApplication;
 import org.store.clothstar.member.domain.MemberRole;
 import org.store.clothstar.member.dto.request.ModifyMemberRequest;
