+++ src/test/java/org/store/clothstar/member/service/SellerServiceApplicationMockUnitTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import org.store.clothstar.member.application.SellerServiceApplication;
 import org.store.clothstar.member.domain.Seller;
 import org.store.clothstar.member.repository.SellerMybatisRepository;
